refactor: reorganize styles and components for better structure and maintainability
