+++ mobile/app/(auth)/login.tsx
@@ -4,7 +4,6 @@ import {
   Text,
   TextInput,
   TouchableOpacity,
-  StyleSheet,
   StatusBar,
   KeyboardAvoidingView,
   Platform,
@@ -13,9 +12,9 @@ import {
   ActivityIndicator,
   SafeAreaView,
 } from 'react-native';
-import api from '../../utils/api';
 import { useRouter } from 'expo-router';
 import { useAuth } from '../../context/AuthContext';
+import { styles } from '../../styles/auth/loginSytles';
 
 const LoginScreen = () => {
   const router = useRouter();
@@ -160,153 +159,6 @@ const LoginScreen = () => {
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#F5F5F5',
-  },
-  keyboardView: {
-    flex: 1,
-  },
-  errorText: {
-    color: '#D93448', 
-    fontSize: 14,
-    textAlign: 'center',
-    marginBottom: 15, 
-    fontWeight: '500',
-  },
-  scrollContent: {
-    flexGrow: 1,
-    justifyContent: 'center',
-    paddingHorizontal: 20,
-  },
-  header: {
-    alignItems: 'center',
-    marginBottom: 40,
-  },
-  logoText: {
-    fontSize: 32,
-    fontWeight: 'bold',
-    color: '#4A4E9D',
-    letterSpacing: 1,
-  },
-  formContainer: {
-    backgroundColor: '#FFFFFF',
-    borderRadius: 20,
-    padding: 30,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 8,
-    elevation: 5,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: 'bold',
-    color: '#333333',
-    marginBottom: 30,
-  },
-  inputContainer: {
-    marginBottom: 20,
-  },
-  label: {
-    fontSize: 16,
-    color: '#666666',
-    marginBottom: 8,
-    fontWeight: '500',
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#E0E0E0',
-    borderRadius: 12,
-    paddingHorizontal: 16,
-    paddingVertical: 14,
-    fontSize: 16,
-    backgroundColor: '#FFFFFF',
-    color: '#333333',
-  },
-  passwordContainer: {
-    position: 'relative',
-  },
-  passwordInput: {
-    paddingRight: 50,
-  },
-  eyeButton: {
-    position: 'absolute',
-    right: 15,
-    top: 15,
-    padding: 5,
-  },
-  eyeText: {
-    fontSize: 18,
-  },
-  continueButton: {
-    backgroundColor: '#4A4E9D',
-    borderRadius: 12,
-    paddingVertical: 16,
-    alignItems: 'center',
-    marginTop: 20,
-    shadowColor: '#4A4E9D',
-    shadowOffset: {
-      width: 0,
-      height: 4,
-    },
-    shadowOpacity: 0.3,
-    shadowRadius: 8,
-    elevation: 8,
-  },
-  buttonDisabled: {
-    opacity: 0.6,
-  },
-  continueButtonText: {
-    color: '#FFFFFF',
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
-  registerContainer: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginTop: 20,
-    marginBottom: 20,
-  },
-  registerText: {
-    color: '#666666',
-    fontSize: 14,
-  },
-  registerLink: {
-    color: '#4A4E9D',
-    fontSize: 14,
-    fontWeight: 'bold',
-  },
-  googleButton: {
-    backgroundColor: '#E8E8E8',
-    borderRadius: 12,
-    paddingVertical: 14,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 10,
-  },
-  googleIcon: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: '#4285F4',
-    backgroundColor: '#FFFFFF',
-    width: 24,
-    height: 24,
-    textAlign: 'center',
-    borderRadius: 12,
-    lineHeight: 24,
-  },
-  googleButtonText: {
-    color: '#666666',
-    fontSize: 16,
-    fontWeight: '500',
-  },
-});
+
 
 export default LoginScreen;
