+++ mobile/styles/indexStyles.ts
@@ -15,7 +15,6 @@ export const styles = StyleSheet.create({
   logo: {
     width: 200,
     height: 200,
-    resizeMode: 'contain',
     marginBottom: 20,
   },
   slogan: {
